requirements.txt
alpaca-trade-api
asyncpg
pandas
numpy
python-dotenv
plotly
streamlit
streamlit-autorefresh
pytz